refactor: move integration tests from main.rs to tests/ directory
- Create lib.rs to expose modules for integration tests
- Add tests/cli_tests.rs for CLI and script discovery tests
- Add tests/key_handling_tests.rs for keyboard event handling tests
- Add tests/app_state_tests.rs for application state tests
- Keep only unit tests for private items in main.rs

This reduces main.rs from ~1026 to ~690 lines, improving maintainability
and following Rust project conventions for test organization.

Closes #27
