+++ src/lib.rs
@@ -1,0 +1,7 @@
+//! Jarvis TUI - A beautiful TUI for managing and executing scripts
+//!
+//! This library provides the core functionality for discovering, parsing,
+//! and executing scripts from various sources (bash, npm, devbox, taskfiles).
+
+pub mod script;
+pub mod ui;
